requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
dbt-postgres==1.8.0
pandas==2.2.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9